requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
langchain>=0.0.350
langchain-community>=0.0.10
langchain-openai>=0.0.2
langchain-core>=0.1.0
python-dotenv>=1.0.0
openai>=1.3.5
requests>=2.31.0
fhir.resources>=7.0.2
pydantic>=1.9.0,<2.0.0
python-multipart>=0.0.6
typing-extensions>=4.5.0
starlette>=0.27.0
anyio>=3.7.1
click>=8.0.0
h11>=0.14.0
httpx>=0.24.1
idna>=3.4
sniffio>=1.3.0
tqdm>=4.65.0
aiohttp>=3.8.5
aiosignal>=1.3.1
async-timeout>=4.0.3
attrs>=23.1.0
frozenlist>=1.4.0
multidict>=6.0.4
yarl>=1.9.2 